+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Improve expand/select handling and support `*` notation
+- Expose CDS OData V2 Adapter singleton instance from CDS plugin at `cds.cov2ap`
+- Allow to register `before` express routes for CDS OData V2 Adapter router via `cds.cov2ap.before`
+- Option to cache and stream generated EDMX from disk instead of keeping it in memory
 
 ## Version 1.12.10 - 2024-06-05